+++ api/src/AppBundle/Controller/ChecklistController.php
@@ -18,50 +18,6 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 class ChecklistController extends RestController
 {
     /**
-     * @Route("/queued", methods={"GET"})
-     * @return array
-     * @throws \Doctrine\DBAL\DBALException
-     */
-    public function getQueuedChecklists(Request $request, EntityManagerInterface $em): array
-    {
-        if (!$this->getAuthService()->isSecretValid($request)) {
-            throw new UnauthorisedException('client secret not accepted.');
-        }
-
-        /** @var array $data */
-        $data = $this->deserializeBodyContent($request);
-
-        /** @var ChecklistRepository $checklistRepo */
-        $checklistRepo = $em->getRepository(Checklist::class);
-
-        $queuedReportIds = $checklistRepo->getReportsIdsWithQueuedChecklistsAndSetChecklistsToInProgress(intval($data['row_limit']));
-
-        $reports = [];
-        foreach ($queuedReportIds as $reportId) {
-            $reports[] = $this->findEntityBy(Report::class, $reportId);
-        }
-
-        $this->setJmsSerialiserGroups([
-            'report-id',
-            'checklist',
-            'user-name',
-            'user-rolename',
-            'report-checklist',
-            'report-sections',
-            'prof-deputy-estimate-management-costs',
-            'checklist-information',
-            'report-client',
-            'report-period',
-            'client-name',
-            'document-sync',
-            'report-submission-uuid',
-            'client-case-number'
-        ]);
-
-        return $reports;
-    }
-
-    /**
      * @Route("/{id}", methods={"PUT"})
      */
     public function update(Request $request, int $id, EntityManagerInterface $em): Checklist
@@ -93,30 +49,6 @@ class ChecklistController extends RestController
 
         if (!empty($data['uuid'])) {
             $checklist->setUuid($data['uuid']);
-        }
-
-        $this->persistAndFlush($checklist);
-
-        $serialisedGroups = ['checklist-id'];
-        $this->setJmsSerialiserGroups($serialisedGroups);
-
-        return $checklist;
-    }
-
-    /**
-     * @Route("/{id}/update-sync-status", methods={"PUT"})
-     * @Security("has_role('ROLE_ADMIN')")
-     */
-    public function updateSyncStatus(Request $request, int $id, EntityManagerInterface $em): Checklist
-    {
-        /** @var array $data */
-        $data = $this->deserializeBodyContent($request);
-
-        /** @var Checklist $checklist */
-        $checklist = $em->getRepository(Checklist::class)->find($id);
-
-        if (!empty($data['syncStatus'])) {
-            $checklist->setSynchronisationStatus($data['syncStatus']);
         }
 
         $this->persistAndFlush($checklist);
