+++ api/tests/AppBundle/Controller/ChecklistControllerTest.php
@@ -1,0 +1,20 @@
+<?php declare(strict_types=1);
+
+namespace Tests\AppBundle\Controller;
+
+class ChecklistControllerTest extends AbstractTestController
+{
+    public function setUp(): void
+    {
+
+    }
+
+
+
+    public static function tearDownAfterClass(): void
+    {
+        parent::tearDownAfterClass();
+
+        self::fixtures()->clear();
+    }
+}
