+++ api/tests/AppBundle/Entity/Repository/ReportRepositoryTest.php
@@ -4,11 +4,17 @@ namespace Tests\AppBundle\Entity\Repository;
 
 use AppBundle\Entity\Client;
 use AppBundle\Entity\ClientInterface;
+use AppBundle\Entity\Report\Checklist;
 use AppBundle\Entity\Report\Fee;
 use AppBundle\Entity\Report\Report;
+use AppBundle\Entity\Report\ReportSubmission;
 use AppBundle\Entity\Repository\ReportRepository;
+use AppBundle\Entity\SynchronisableInterface;
 use AppBundle\Entity\User;
+use DateInterval;
+use DateTimeZone;
 use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Mapping\ClassMetadata;
 use Mockery\MockInterface;
@@ -18,12 +24,6 @@ use Mockery as m;
 
 class ReportRepositoryTest extends WebTestCase
 {
-    /**
-     * @var Fixtures|null
-     */
-    private static $fixtures;
-    private static $repo;
-
     /**
      * @var ReportRepository
      */
@@ -44,6 +44,18 @@ class ReportRepositoryTest extends WebTestCase
 
     private $mockMetaClass;
 
+    /** @var ReportRepository */
+    private $repository;
+
+    /** @var EntityManager */
+    private $entityManager;
+
+    /** @var array */
+    private $queryResult;
+
+    /** @var int */
+    const QUERY_LIMIT = 2;
+
     public function setUp(): void
     {
         $this->mockEm = m::mock(EntityManagerInterface::class);
@@ -56,6 +68,11 @@ class ReportRepositoryTest extends WebTestCase
             ->andReturn($this->mockClient);
 
         $this->sut = new ReportRepository($this->mockEm, $this->mockMetaClass);
+
+        $kernel = self::bootKernel();
+
+        $this->entityManager = $kernel->getContainer()->get('doctrine')->getManager();
+        $this->repository = $this->entityManager->getRepository(Report::class);
     }
 
     public function testAddFeesToReportIfMissingForNonPAUser()
@@ -90,5 +107,128 @@ class ReportRepositoryTest extends WebTestCase
         $this->mockReport->shouldReceive('isPAReport')->andReturn(true);
 
         $this->assertEquals(0, $this->sut->addFeesToReportIfMissing($this->mockReport));
+    }
+
+    /**
+     * @test
+     */
+    public function fetchQueuedChecklists()
+    {
+        $this
+            ->ensureChecklistsExistInDatabase()
+            ->fetchChecklists()
+            ->assertOnlyAlimitedNumberOfQueuedChecklistsAreReturned()
+            ->assertQueuedChecklistsAreUpdatedToInProgress();
+    }
+
+    /**
+     * @return ReportRepositoryTest
+     * @throws \Doctrine\ORM\ORMException
+     * @throws \Doctrine\ORM\OptimisticLockException
+     */
+    private function ensureChecklistsExistInDatabase(): ReportRepositoryTest
+    {
+        $client = (new Client())->setCaseNumber('49329657');
+        $this->entityManager->persist($client);
+
+        $this
+            ->buildChecklistWithStatus($client, SynchronisableInterface::SYNC_STATUS_QUEUED)
+            ->buildChecklistWithStatus($client, SynchronisableInterface::SYNC_STATUS_QUEUED)
+            ->buildChecklistWithStatus($client, SynchronisableInterface::SYNC_STATUS_QUEUED)
+            ->buildChecklistWithStatus($client, SynchronisableInterface::SYNC_STATUS_SUCCESS)
+            ->buildChecklistWithStatus($client, null);
+
+        $this->entityManager->flush();
+
+        return $this;
+    }
+
+    /**
+     * @return ReportRepositoryTest
+     * @throws \Doctrine\DBAL\DBALException
+     */
+    private function fetchChecklists(): ReportRepositoryTest
+    {
+        $this->queryResult = $this->repository->getReportsIdsWithQueuedChecklistsAndSetChecklistsToInProgress(self::QUERY_LIMIT);
+        return $this;
+    }
+
+    /**
+     * @return ReportRepositoryTest
+     */
+    private function assertOnlyAlimitedNumberOfQueuedChecklistsAreReturned(): ReportRepositoryTest
+    {
+        $this->assertCount(self::QUERY_LIMIT, $this->queryResult);
+        $this->assertEquals(1, $this->queryResult[0]);
+        $this->assertEquals(2, $this->queryResult[1]);
+
+        return $this;
+    }
+
+    private function assertQueuedChecklistsAreUpdatedToInProgress(): void
+    {
+        $repository = $this->entityManager->getRepository(Checklist::class);
+        $result = $repository->findBy(['synchronisationStatus' => SynchronisableInterface::SYNC_STATUS_IN_PROGRESS]);
+        $this->assertCount(2, $result);
+        $this->assertEquals(1, $result[0]->getId());
+        $this->assertEquals(2, $result[1]->getId());
+    }
+
+    /**
+     * @param Client $client
+     * @param string|null $status
+     * @return ReportRepositoryTest
+     * @throws \Doctrine\ORM\ORMException
+     */
+    private function buildChecklistWithStatus(Client $client, ?string $status): ReportRepositoryTest
+    {
+        $report = $this->buildReport($client);
+        $checklist = new Checklist($report);
+
+        if ($status) {
+            $checklist->setSynchronisationStatus($status);
+        }
+
+        $this->entityManager->persist($checklist);
+
+        return $this;
+    }
+
+    /**
+     * @param Client $client
+     * @return Report
+     * @throws \Exception
+     */
+    private function buildReport(Client $client): Report
+    {
+        $startDate = new \DateTime('now', new DateTimeZone('UTC'));
+        $endDate = $startDate->add(new DateInterval('P1D'));
+        $report = new Report($client, Report::TYPE_PROPERTY_AND_AFFAIRS_HIGH_ASSETS, $startDate, $endDate);
+
+        $user = (new User())
+            ->setFirstname('firstname')
+            ->setLastname('lastname')
+            ->setEmail(sprintf('email%s@test.com', rand(1, 1000)))
+            ->setPassword('password');
+
+        $reportSubmission = new ReportSubmission($report, $user);
+        $report->addReportSubmission($reportSubmission);
+
+        $this->entityManager->persist($report);
+        $this->entityManager->persist($reportSubmission);
+        $this->entityManager->persist($user);
+
+        return $report;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    protected function tearDown(): void
+    {
+        parent::tearDown();
+
+        $this->entityManager->close();
+        $this->entityManager = null;
     }
 }
